+++ backend/billboard/src/main/java/com/ssafy/billboard/model/repository/BoardGameRepository.java
@@ -3,6 +3,6 @@ package com.ssafy.billboard.model.repository;
 import com.ssafy.billboard.model.entity.BoardGame;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardGameRepository extends JpaRepository<BoardGame,Long> {
+public interface BoardGameRepository extends JpaRepository<BoardGame,Integer> {
 
 }
